re-structured files in templates
