+++ js/templates/creators/index.mjs
@@ -1,0 +1,1 @@
+export * from "./postHtml.mjs";
