+++ js/templates/render/index.mjs
@@ -1,3 +1,2 @@
 export * from "./post.mjs";
 export * from "./posts.mjs";
-export * from "./profile.mjs";
