+++ js/templates/render/post.mjs
@@ -1,0 +1,5 @@
+import * as template from "../creators/index.mjs";
+
+export function post(postData, parent) {
+  parent.append(template.postTemplate(postData));
+}
