+++ js/templates/render/posts.mjs
@@ -1,0 +1,5 @@
+import * as template from "../creators/index.mjs";
+
+export function posts(postDataList, parent) {
+  parent.append(...postDataList.map(template.postTemplate));
+}
